# logging: Add traceID field to access logs when tracing is active