@francislavoie 2023-04-23
Related to https://github.com/caddyserver/caddy/issues/5336