@francislavoie 2023-04-23
I decided to piggy-back off this PR for https://github.com/caddyserver/caddy/issues/5336.

I don't love this though. This is kinda going into #4649 territory, we're doing allocations for the logging. We could use a `sync.Pool` for `ExtraLogFields` but idk 🤔 pooling isn't perfect, long-running requests/connections like websockets will hold onto them for a long time etc. We can throw away my commits if the approach doesn't make sense.

I probably want to move `user_id` to only log if non-empty (as an extra field essentially), and later on we might want user-configurable extra log fields for things like `http.request.uuid` or whatever.